Extract Results interface (from ResultsAbstract).
