+++ engine/src/main/java/nl/inl/blacklab/search/results/package-info.java
@@ -3,7 +3,7 @@
  *
  * Classes for dealing with search results: {@link nl.inl.blacklab.search.results.Hits},
  * {@link nl.inl.blacklab.search.results.DocResults}, {@link nl.inl.blacklab.search.results.HitGroups},
- * {@link nl.inl.blacklab.search.results.DocGroups}. All derive from {@link nl.inl.blacklab.search.results.Results}.
+ * {@link nl.inl.blacklab.search.results.DocGroups}. All derive from {@link nl.inl.blacklab.search.results.ResultsAbstract}.
  *
  * Sorting and grouping can be done with {@link nl.inl.blacklab.resultproperty.HitProperty},
  * {@link nl.inl.blacklab.resultproperty.DocProperty}, etc.
