+++ engine/src/main/java/nl/inl/blacklab/searches/SearchCountFromResults.java
@@ -6,7 +6,7 @@ import nl.inl.blacklab.exceptions.InvalidQuery;
 import nl.inl.blacklab.search.results.QueryInfo;
 import nl.inl.blacklab.search.results.ResultCount;
 import nl.inl.blacklab.search.results.ResultCount.CountType;
-import nl.inl.blacklab.search.results.Results;
+import nl.inl.blacklab.search.results.ResultsAbstract;
 import nl.inl.blacklab.search.results.ResultsStats;
 import nl.inl.blacklab.search.results.ResultsStatsDelegate;
 
@@ -14,7 +14,7 @@ import nl.inl.blacklab.search.results.ResultsStatsDelegate;
  * A search operation that yields a count as its result.
  * @param <T> result type, e.g. Hit
  */
-public class SearchCountFromResults<T extends Results<?, ?>> extends SearchCount {
+public class SearchCountFromResults<T extends ResultsAbstract<?, ?>> extends SearchCount {
 
     /**
      * The search we're doing a count for.
