+++ engine/src/main/java/nl/inl/blacklab/searches/SearchForResults.java
@@ -2,12 +2,12 @@ package nl.inl.blacklab.searches;
 
 import nl.inl.blacklab.search.results.QueryInfo;
 import nl.inl.blacklab.search.results.ResultCount.CountType;
-import nl.inl.blacklab.search.results.Results;
+import nl.inl.blacklab.search.results.ResultsAbstract;
 
 /** A search that yields results.
  * @param <R> results type, e.g. Hits
  */
-public abstract class SearchForResults<R extends Results<?, ?>> extends AbstractSearch<R> {
+public abstract class SearchForResults<R extends ResultsAbstract<?, ?>> extends AbstractSearch<R> {
 
     SearchForResults(QueryInfo queryInfo) {
         super(queryInfo);
